Issue #11534: Issue #11542: Publications test structure changes
