+++ profile/modules/apps/os_publications/tests/src/ExistingSite/PublicationsViewsTest.php
@@ -2,9 +2,6 @@
 
 namespace Drupal\Tests\os_publications\ExistingSite;
 
-use Drupal\Component\Render\FormattableMarkup;
-use Drupal\views\Plugin\views\field\EntityField;
-use Drupal\views\ViewExecutable;
 use Drupal\views\Views;
 
 /**
@@ -54,7 +51,7 @@ class PublicationsViewsTest extends TestBase {
    *
    * @throws \Drupal\Core\Entity\EntityStorageException
    */
-  public function testType() {
+  public function testType(): void {
     $this->createReference([
       'title' => 'Mona Lisa',
       'is_sticky' => [
@@ -93,7 +90,7 @@ class PublicationsViewsTest extends TestBase {
    *
    * @throws \Drupal\Core\Entity\EntityStorageException
    */
-  public function testTitle() {
+  public function testTitle(): void {
     $this->createReference([
       'title' => 'The Last Supper',
       'is_sticky' => [
@@ -148,7 +145,7 @@ class PublicationsViewsTest extends TestBase {
    * @throws \Drupal\Core\Entity\EntityStorageException
    * @throws \Drupal\Core\TypedData\Exception\MissingDataException
    */
-  public function testAuthor() {
+  public function testAuthor(): void {
     /** @var \Drupal\os_publications\PublicationsListingHelperInterface $publications_listing_helper */
     $publications_listing_helper = $this->container->get('os_publications.listing_helper');
 
@@ -329,7 +326,7 @@ class PublicationsViewsTest extends TestBase {
    *
    * @throws \Drupal\Core\Entity\EntityStorageException
    */
-  public function testYear() {
+  public function testYear(): void {
     $this->createReference([
       'title' => 'Girl with a Pearl Earring',
       'bibcite_year' => [
@@ -393,7 +390,7 @@ class PublicationsViewsTest extends TestBase {
    *
    * @throws \Drupal\Core\Entity\EntityStorageException
    */
-  public function testReferenceFilter() {
+  public function testReferenceFilter(): void {
     $this->createReference([
       'title' => 'Girl with a Pearl Earring',
       'is_sticky' => [
@@ -449,7 +446,7 @@ class PublicationsViewsTest extends TestBase {
    * @throws \Drupal\Core\Entity\EntityStorageException
    * @throws \Drupal\Core\TypedData\Exception\MissingDataException
    */
-  public function testSortOrder() {
+  public function testSortOrder(): void {
     $reference1 = $this->createReference([
       'title' => 'Girl with a Pearl Earring',
       'bibcite_year' => [
@@ -545,132 +542,6 @@ class PublicationsViewsTest extends TestBase {
     $this->assertIdenticalResultset($view, $dataset, [
       '_entity' => 'title',
     ]);
-  }
-
-  /**
-   * Orders a nested array containing a result set based on a given column.
-   *
-   * Copied from ViewsKernelTestBase::orderResultSet.
-   *
-   * @param array $result_set
-   *   An array of rows from a result set, with each row as an associative
-   *   array keyed by column name.
-   * @param string $column
-   *   The column name by which to sort the result set.
-   * @param bool $reverse
-   *   (optional) Boolean indicating whether to sort the result set in reverse
-   *   order. Defaults to FALSE.
-   *
-   * @see \Drupal\Tests\views\Kernel\ViewsKernelTestBase::orderResultSet
-   */
-  protected function orderResultSet(array &$result_set, $column, $reverse = FALSE) {
-    $order = $reverse ? -1 : 1;
-    usort($result_set, function ($a, $b) use ($column, $order) {
-      if ($a[$column] === $b[$column]) {
-        return 0;
-      }
-      return $order * (($a[$column] < $b[$column]) ? -1 : 1);
-    });
-  }
-
-  /**
-   * Verifies that a result set returned by a View matches expected values.
-   *
-   * The comparison is done on the string representation of the columns of the
-   * column map, taking the order of the rows into account, but not the order
-   * of the columns.
-   *
-   * Copied from ViewResultAssertionTrait::assertIdenticalResultset.
-   *
-   * @param \Drupal\views\ViewExecutable $view
-   *   An executed View.
-   * @param array $expected_result
-   *   An expected result set.
-   * @param array $column_map
-   *   (optional) An associative array mapping the columns of the result set
-   *   from the view (as keys) and the expected result set (as values).
-   * @param string $message
-   *   (optional) A custom message to display with the assertion. Defaults to
-   *   'Identical result set.'.
-   *
-   * @see \Drupal\views\Tests\ViewResultAssertionTrait::assertIdenticalResultset
-   */
-  protected function assertIdenticalResultset(ViewExecutable $view, array $expected_result, array $column_map = [], $message = NULL) {
-    // Convert $view->result to an array of arrays.
-    $result = [];
-    foreach ($view->result as $key => $value) {
-      $row = [];
-      foreach ($column_map as $view_column => $expected_column) {
-        if (property_exists($value, $view_column)) {
-          $row[$expected_column] = (string) $value->$view_column->label();
-        }
-        // For entity fields we don't have the raw value. Let's try to fetch it
-        // using the entity itself.
-        elseif (empty($value->$view_column) && isset($view->field[$expected_column]) && ($field = $view->field[$expected_column]) && $field instanceof EntityField) {
-          $column = NULL;
-          if (count(explode(':', $view_column)) == 2) {
-            $column = explode(':', $view_column)[1];
-          }
-          // The comparison will be done on the string representation of the
-          // value.
-          $field_value = $field->getValue($value, $column);
-          $row[$expected_column] = is_array($field_value) ? array_map('strval', $field_value) : (string) $field_value;
-        }
-      }
-      $result[$key] = $row;
-    }
-
-    // Remove the columns we don't need from the expected result.
-    foreach ($expected_result as $key => $value) {
-      $row = [];
-      foreach ($column_map as $expected_column) {
-        // The comparison will be done on the string representation of the
-        // value.
-        if (is_object($value)) {
-          $row[$expected_column] = (string) $value->$expected_column;
-        }
-        // This case is about fields with multiple values.
-        elseif (is_array($value[$expected_column])) {
-          foreach (array_keys($value[$expected_column]) as $delta) {
-            $row[$expected_column][$delta] = (string) $value[$expected_column][$delta];
-          }
-        }
-        else {
-          $row[$expected_column] = (string) $value[$expected_column];
-        }
-      }
-      $expected_result[$key] = $row;
-    }
-
-    $this->verbose('<pre style="white-space: pre-wrap;">'
-      . "\n\nQuery:\n" . $view->build_info['query']
-      . "\n\nQuery arguments:\n" . var_export($view->build_info['query']->getArguments(), TRUE)
-      . "\n\nActual result:\n" . var_export($result, TRUE)
-      . "\n\nExpected result:\n" . var_export($expected_result, TRUE));
-
-    // Reset the numbering of the arrays.
-    $result = array_values($result);
-    $expected_result = array_values($expected_result);
-
-    // Do the actual comparison.
-    if (!isset($message)) {
-      $message = new FormattableMarkup("Actual result <pre>\n@actual\n</pre> is not identical to expected <pre>\n@expected\n</pre>", [
-        '@actual' => var_export($result, TRUE),
-        '@expected' => var_export($expected_result, TRUE),
-      ]);
-    }
-    return $this->assertSame($result, $expected_result, $message);
-  }
-
-  /**
-   * Check anonymous user access to publications.
-   *
-   * @throws \Behat\Mink\Exception\ExpectationException
-   */
-  public function testAnonymousUserAccess() {
-    $this->visit('/publications');
-
-    $this->assertSession()->statusCodeEquals(403);
   }
 
   /**
